requirements.txt
aiogram>=3.0.0
httpx
python-dotenv
aiohttp